% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{ApplyNormalization}
\alias{ApplyNormalization}
\title{Applies normalization on sequencing data}
\usage{
ApplyNormalization(
  data,
  method = c("CPM", "UQ", "TMM", "DESeq", "RLE", "RUV"),
  control.idx = NULL,
  sc.idx = NULL
)
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is the number of samples and p is the number of features.}

\item{method}{Vector of normalization methods that are applied to the data.
Available methods are: \code{c("CPM", "UQ", "TMM", "DESeq", "RUV")}.
Select one or multiple methods. By default all normalization methods will be applied.}

\item{control.idx}{Vector of control genes' id.}

\item{sc.idx}{A numeric matrix specifying the replicate samples for which to
compute the count differences used to estimate the factors of unwanted variation.}
}
\value{
List of objects containing normalized data and associated normalization factors.
}
\description{
Applies normalization on sequencing data
}
