% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{AssessDiffAnalysis}
\alias{AssessDiffAnalysis}
\title{Assess differential analysis performance}
\usage{
AssessDiffAnalysis(de.raw.res, de.norm.res)
}
\arguments{
\item{de.raw.res}{Raw counts DE results}

\item{de.norm.res}{Normalized counts DE results}
}
\value{
List of similarity
}
\description{
Assess differential analysis performance
}
