% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{AssessNormalization}
\alias{AssessNormalization}
\title{Assessment of normalization}
\usage{
AssessNormalization(data.raw, data.normalized, pos.controls, enrich.idx)
}
\arguments{
\item{data.raw}{A un-normalized count data matrix of shape n x p, where n is
the number of samples and p is the number of features.}

\item{data.normalized}{A normalized count data matrix (n x p).}

\item{pos.controls}{Vector of positive enriched genes.}

\item{enrich.idx}{Matrix with two rows indicating the column index of
enrichment and input samples in the raw/normalized count data matrix.
The first row is the column index of input and the second row is the
column index of enrichment samples.}

\item{de.raw.res}{List of un-normalized DE results}

\item{de.norm.res}{List of normalized DE results}
}
\value{
List of metrics that assessing normalization effect.
}
\description{
Assessment of normalization
}
