% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{BetweenStatPlot}
\alias{BetweenStatPlot}
\title{Box-violin plot comparing values between groups}
\usage{
BetweenStatPlot(
  data,
  x,
  y,
  color,
  palette = NULL,
  test = c("wilcox.test", "t.test", "none")
)
}
\arguments{
\item{data}{A data frame (or a tibble).}

\item{x}{The grouping variable from the \code{data}.}

\item{y}{The value variable from the \code{data}.}

\item{color}{The color variable from the \code{data}.}

\item{palette}{The color palette for different groups.}

\item{test}{Perform "wilcox.test" or "t.test" or not test.}
}
\value{
ggplot2 object
}
\description{
Box-violin plot comparing values between groups
}
