% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calcCorDiff}
\alias{calcCorDiff}
\title{Calculate difference of correlation}
\usage{
calcCorDiff(raw.cor, norm.cor)
}
\arguments{
\item{raw.cor}{Correlation matrix (p x p) computed from raw count matrix, where p is the number of samples.}

\item{norm.cor}{Correlation matrix (p x p) computed from normalized count matrix, where p is the number of samples.}
}
\value{

}
\description{
Calculate difference of correlation
}
