% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calcSizeFactors}
\alias{calcSizeFactors}
\title{Calculate size factors for scaling raw library size}
\usage{
calcSizeFactors(data, method = NULL, control.prefix = NULL)
}
\arguments{
\item{data}{A counts matrix.}

\item{method}{Normalization methods to be used. Should be one of c("TMM","TMMwsp","RLE","upperquartile","none","spikein").}

\item{control.prefix}{Prefix of control gene id.}
}
\value{
values of size factors
}
\description{
Calculate size factors for scaling raw library size
}
