% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Enone-class}
\alias{Enone-class}
\alias{createEnone}
\title{Enone object and constructor}
\usage{
createEnone(
  data,
  bio.group,
  enrich.group,
  batch.group = NULL,
  spike.in.prefix,
  input.id = "Input",
  enrich.id = "Enrich",
  synthetic.id = NULL
)
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is the
number of samples and p is the number of features.}

\item{bio.group}{Vector of samples group, e.g., c("Young.Input","Young.Enrich","Old.Input","Old.Enrich").}

\item{enrich.group}{Vector of enrichment group, e.g., c("Input","Enrich","Input","Enrich").}

\item{batch.group}{Vector of samples batch, e.g., c("A","A","B","B"), default=NULL.}

\item{spike.in.prefix}{A character specify the prefix of spike-in id, e.g., "FB" stands for fly spike-in id.}

\item{input.id}{Input library id, must be consistent with \code{enrich.group}, e.g., "Input".}

\item{enrich.id}{Enrich library id, must be consistent with \code{enrich.group}, e.g., "Enrich".}

\item{synthetic.id}{Vector of synthetic RNA id, e.g. c("Syn1","Syn2"), default=NULL.

Description of each slot:
\code{assay} \code{SummarizedExperiment::Assays} object, contains all counts.
\code{counts} list.
\code{enone_factor} list.
\code{enone_metrics} data.frame.
\code{enone_score} data.frame.
\code{enrichment} list.
\code{enrichment_filtered} list.
\code{parameter} list.}
}
\value{
A Enone S4 object
}
\description{
\code{Enone} object extends the \code{SummarizedExperiment} class.
The \code{createEnone} is a easy constructor of \code{Enone} object
}
