% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DESeq2DE}
\alias{DESeq2DE}
\title{Wrapper of DESeq2 procedure}
\usage{
DESeq2DE(counts, group, design.formula, contrast.df)
}
\arguments{
\item{counts}{A un-normalized counts data matrix.}

\item{group}{Vector of length p mapping the columns of \code{counts} to
corresponding samples group.}

\item{design.formula}{Formula}

\item{contrast.df}{Data frame of contrast, where extracting results as
first column vs. second column.}
}
\value{
List containing differential analysis object, result table and filtered result table.
}
\description{
Wrapper of DESeq2 procedure
}
