% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{DiffAnalysis}
\alias{DiffAnalysis}
\title{Applying Differential analysis}
\usage{
DiffAnalysis(data.raw, data.norm.ls, group)
}
\arguments{
\item{data.raw}{A un-normalized count data matrix of shape n x p, where n is
the number of samples and p is the number of features.}

\item{data.norm.ls}{List containing normalized counts and adjust factors for
adjusting unwanted variation.}

\item{group}{Vector of length p mapping the columns of \code{data.raw} to
corresponding samples group.}
}
\value{
List containing differential analysis object, result table and filtered result table.
}
\description{
Applying Differential analysis
}
