% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{edgeRDE}
\alias{edgeRDE}
\title{Wrapper of edgeR procedure}
\usage{
edgeRDE(
  counts,
  group,
  norm.factors = NULL,
  adjust.factors = NULL,
  design.formula = NULL,
  contrast.df = NULL,
  coef = NULL
)
}
\arguments{
\item{counts}{A un-normalized counts data matrix.}

\item{group}{Vector of length p mapping the columns of \code{counts} to
corresponding samples group.}

\item{norm.factors}{Vector of normalization factors with p length.}

\item{adjust.factors}{Matrix with each column indicates the adjusting factors
that estimated from RUV.}

\item{design.formula}{Formula}

\item{contrast.df}{Data frame of contrast, where extracting results as
first column vs. second column.}

\item{coef}{integer or character vector indicating which coefficients of the
linear model are to be tested equal to zero. Values must be columns or column names of design.}
}
\value{
List containing differential analysis object, result table and filtered result table.
}
\description{
Wrapper of edgeR procedure
}
