% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{enONE}
\alias{enONE}
\title{Normalization and normalization assessment in one function}
\usage{
enONE(
  data,
  group,
  spike.in.prefix,
  input.id = "Input",
  enrich.id = "Enrich",
  n.neg.control = 1000,
  n.pos.eval = 1000,
  n.neg.eval = 1000,
  scaling.method = c("TC", "UQ", "TMM", "DESeq"),
  ruv.norm = TRUE,
  ruv.k = 1,
  ruv.drop = 0,
  pam_krange = 2:6,
  pc_k = 3
)
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is the number of samples and p is the number of features.}

\item{group}{Vector of samples group, e.g., c("Young.Input","Young.Enrich")}

\item{spike.in.prefix}{A character specify the prefix of spike-in id.}

\item{input.id}{Input library id, must be consistent with \code{group}, e.g., "Input".}

\item{enrich.id}{Enrich library id, must be consistent with \code{group}, e.g., "Enrich".}

\item{n.neg.control}{Number of negative control genes for RUV normalization.}

\item{n.pos.eval}{Number of positive evaluation genes for wanted variation assessment.}

\item{n.neg.eval}{Number of negative evaluation genes for unwanted variation assessment.}

\item{scaling.method}{Vector of normalization methods that are applied to the data.
Available methods are: \code{c("TC", "UQ", "TMM", "DESeq")}.
Select one or multiple methods. By default all normalization methods will be applied.}

\item{ruv.norm}{Whether to perform RUV normalization.}

\item{ruv.k}{The number of factors of unwanted variation to be estimated from the data.}

\item{ruv.drop}{The number of singular values to drop in the estimation of
unwanted variation, default drop the first singular value that represent the
difference between enrichment and input.}

\item{pam_krange}{Integer or vector of integers indicates the number of
clusters for PAM clustering, default: 2:6.}

\item{pc_k}{Integer indicates the metrics will be calculated in the first kth PCs, default: 3.}
}
\value{
list
}
\description{
Normalization and normalization assessment in one function
}
