% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{FindEnrichment}
\alias{FindEnrichment}
\alias{FindEnrichment,Enone,character,character,numeric,numeric-method}
\title{Find enriched genes between enrich and input samples}
\usage{
FindEnrichment(object, slot, method, fc.cutoff, p.cutoff)

\S4method{FindEnrichment}{Enone,character,character,numeric,numeric}(
  object,
  slot = c("sample", "spike_in"),
  method,
  fc.cutoff = 1,
  p.cutoff = 0.05
)
}
\arguments{
\item{object}{Enone.}

\item{slot}{Which slot to get, one of \code{sample} or \code{spike_in}.}

\item{method}{Which normalization methods to get, must be one of the methods presented in the selected slot.}

\item{fc.cutoff}{Integer indicate the cutoff for log2-Fold-Change, only
consider genes with log2-Fold-Change greater than this cutoff, default: 1.}

\item{p.cutoff}{Numeric indicate the cutoff for adjusted p-value (e.g., FDR),
genes with adjusted p-value smaller than cutoff were consider, default: 0.05.}
}
\value{
data.frame
}
\description{
Find enriched genes between enrich and input samples
}
