% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{getFactor}
\alias{getFactor}
\alias{getFactor,Enone,character,character,character-method}
\title{Accessor of enONE normalization factors}
\usage{
getFactor(object, slot, method, which)

\S4method{getFactor}{Enone,character,character,character}(object, slot, method, which)
}
\arguments{
\item{object}{Enone.}

\item{slot}{character.}

\item{method}{character.}

\item{which}{character.}
}
\value{
vector or list of factors.
}
\description{
Accessor of enONE normalization factors
}
