% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ggPCA}
\alias{ggPCA}
\title{PCA plot from counts matrix}
\usage{
ggPCA(object, labels = colnames(object), vst.norm = TRUE)
}
\arguments{
\item{object}{A count matrix}

\item{labels}{character vector of sample names or labels. Defaults to colnames(object).}

\item{vst.norm}{if TRUE perform vst transformation.}
}
\value{
ggplot2 object
}
\description{
PCA plot from counts matrix
}
