% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ggPCA_Biplot}
\alias{ggPCA_Biplot}
\title{Biplot of individuals and variables}
\usage{
ggPCA_Biplot(object, performance_score, ...)
}
\arguments{
\item{object}{PCA object returned by \code{prcomp}.}

\item{performance_score}{Vector of performance scores from ass.}

\item{...}{Additional parameters can be passed to \code{factoextra::fviz_pca_biplot()}.}
}
\value{
ggplot2 object
}
\description{
Biplot of individuals and variables
}
