% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{JacIdx}
\alias{JacIdx}
\title{Calculate pairwise set similarity using jaccard index}
\usage{
JacIdx(x, y)
}
\arguments{
\item{x}{vector of set 1.}

\item{y}{vector of set 2.}
}
\value{
pairwise similarity
}
\description{
Calculate pairwise set similarity using jaccard index
}
