% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{JCPlot}
\alias{JCPlot}
\title{Visualization of average jaccard index in different k}
\usage{
JCPlot(k.cor.vec, ref.cor = NULL)
}
\arguments{
\item{k.cor.vec}{vector of overall similarity in different k.}

\item{ref.cor}{value of reference similarity.}
}
\value{
ggplot2 object
}
\description{
Visualization of average jaccard index in different k
}
