% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{normRUV}
\alias{normRUV}
\title{Perform RUV normalization}
\usage{
normRUV(data, control.idx = NULL, sc.idx = NULL, method = c("RUVg", "RUVs"))
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is
the number of samples and p is the number of features.}

\item{control.idx}{Vector of control genes' id.}

\item{sc.idx}{A numeric matrix specifying the replicate samples for which to
compute the count differences used to estimate the factors of unwanted variation.}

\item{method}{Perform RUVg or RUVs normalization.}
}
\value{
List containing normalized counts and adjust factors for adjusting unwanted variation.
}
\description{
Perform RUV normalization
}
