% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{reduceRes}
\alias{reduceRes}
\title{Combine list of DE results}
\usage{
reduceRes(res.ls, fc.col, levels = names(res.ls))
}
\arguments{
\item{res.ls}{Named list of DE results. Each elements in the list correspond to a table of DE results between groups.}

\item{fc.col}{Name of the fold change column.}

\item{levels}{Factor levels of the DE group.}
}
\value{
data.frame
}
\description{
Combine list of DE results
}
