% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setSimilarity}
\alias{setSimilarity}
\title{Calculate pairwise set similarity using jaccard index}
\usage{
setSimilarity(set.ls, pair = FALSE)
}
\arguments{
\item{set.ls}{list of sets}

\item{pair}{If TRUE return pairwise similarity, otherwise return overall similarity (average).}
}
\value{
Vector of similarity
}
\description{
Calculate pairwise set similarity using jaccard index
}
