% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{SynFC}
\alias{SynFC}
\title{Calculate fold-change of synthetic RNA}
\usage{
SynFC(dat.norm.ls, syn.id, enrich.idx)
}
\arguments{
\item{dat.norm.ls}{List containing normalized counts and adjust factors for
adjusting unwanted variation.}

\item{syn.id}{Vector of synthetic RNA ids.}

\item{enrich.idx}{Matrix with two rows indicating the column index of
enrichment and input samples in the raw/normalized count data matrix.
The first row is the column index of input and the second row is the
column index of enrichment samples.}
}
\value{
Data frame with fold-change of synthetic RNA
}
\description{
Calculate fold-change of synthetic RNA
}
