% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{edgeRDE}
\alias{edgeRDE}
\title{Wrapper of edgeR procedure}
\usage{
edgeRDE(
  counts,
  group,
  norm.factors = NULL,
  adjust.factors = NULL,
  design.formula = NULL,
  contrast.df = NULL,
  coef = NULL,
  fc.cutoff = 1,
  p.cutoff = 0.05
)
}
\arguments{
\item{counts}{A un-normalized counts data matrix.}

\item{group}{Vector of length p mapping the columns of \code{counts} to
corresponding samples group.}

\item{norm.factors}{Vector of normalization factors with p length.}

\item{adjust.factors}{Matrix with each column indicates the adjusting factors
that estimated from RUV.}

\item{design.formula}{Formula}

\item{contrast.df}{Data frame of contrast, where extracting results as
first column vs. second column.}

\item{coef}{Integer or character vector indicating which coefficients of the
linear model are to be tested equal to zero. Values must be columns or column names of design.}

\item{fc.cutoff}{Integer indicate the cutoff for log2-Fold-Change, only
consider genes with log2-Fold-Change greater than this cutoff, default: 1.}

\item{p.cutoff}{Numeric indicate the cutoff for adjusted p-value (e.g., FDR),
genes with adjusted p-value smaller than cutoff were consider, default: 0.05.}
}
\value{
List containing differential analysis object, result table and filtered result table.
}
\description{
Wrapper of edgeR procedure
}
