% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{enONE}
\alias{enONE}
\title{Normalization and assessment in one function}
\usage{
enONE(
  object,
  n.neg.control = 1000,
  n.pos.eval = 1000,
  n.neg.eval = 1000,
  scaling.method = c("TC", "UQ", "TMM", "DESeq"),
  ruv.norm = TRUE,
  ruv.k = 1,
  ruv.drop = 0,
  pam.krange = 2:6,
  pc.k = 3
)
}
\arguments{
\item{object}{Enone object.}

\item{n.neg.control}{Number of negative control genes for RUV normalization.}

\item{n.pos.eval}{Number of positive evaluation genes for wanted variation assessment.}

\item{n.neg.eval}{Number of negative evaluation genes for unwanted variation assessment.}

\item{scaling.method}{Vector of normalization methods that are applied to the data.
Available methods are: \code{c("TC", "UQ", "TMM", "DESeq")}.
Select one or multiple methods. By default all normalization methods will be applied.}

\item{ruv.norm}{Whether to perform RUV normalization.}

\item{ruv.k}{The number of factors of unwanted variation to be estimated from the data.}

\item{ruv.drop}{The number of singular values to drop in the estimation of
unwanted variation, default not drop.}

\item{pam.krange}{Integer or vector of integers indicates the number of
clusters for PAM clustering, default: 2:6.}

\item{pc.k}{Integer indicates the metrics will be calculated in the first kth PCs, default: 3.}
}
\value{
list
}
\description{
Normalization and assessment in one function
}
